% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chol_solve}
\alias{chol_solve}
\title{Solve function with Cholesky decomposition.}
\usage{
chol_solve(M, V)
}
\arguments{
\item{M}{A (M x M) symmetric positive definite matrix.}

\item{V}{A (M x 1) vector.}
}
\value{
The solution to M*U = V.
}
\description{
An Rcpp function that solves M*U = V.
}
