% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{concentrated}
\alias{concentrated}
\title{Example data set}
\format{
An object of class \code{list} of length 15.
}
\usage{
concentrated
}
\description{
Contains a list with data and parameters to run the package examples.
Please see \code{?gigg_fixed} and \code{?grouped_igg_mmle} pages for use.
}
\examples{
concentrated
names(concentrated)
}
\keyword{datasets}
