% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{digamma_inv}
\alias{digamma_inv}
\title{Inverse digamma function.}
\usage{
digamma_inv(y, precision = 1e-08)
}
\arguments{
\item{y}{value to evaluate the inverse digamma function at.}

\item{precision}{default = 1e-08.}
}
\description{
Evaluate the inverse diagmma function.
}
