% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gigg_fixed.R
\name{gigg_fixed}
\alias{gigg_fixed}
\title{GIGG regression Gibbs sampler with fixed hyperparameters.}
\usage{
gigg_fixed(
  X,
  C,
  Y,
  grp_idx,
  alpha_inits = rep(0, ncol(C)),
  beta_inits = rep(0, ncol(X)),
  lambda_sq_inits = rep(1, ncol(X)),
  gamma_sq_inits = rep(1, length(unique(grp_idx))),
  a = rep(0.5, length(unique(grp_idx))),
  b = rep(0.5, length(unique(grp_idx))),
  tau_sq_init = 1,
  sigma_sq_init = 1,
  nu_init = 1,
  n_burn_in = 500,
  n_samples = 1000,
  n_thin = 1,
  stable_const = 1e-07,
  verbose = TRUE,
  btrick = FALSE,
  stable_solve = FALSE
)
}
\arguments{
\item{X}{A (n x p) matrix of covariates that we want to apply GIGG shrinkage on.}

\item{C}{A (n x k) matrix of covariates that we want to apply no shrinkage on (typically intercept + adjustment covariates).}

\item{Y}{A (n x 1) column vector of responses.}

\item{grp_idx}{A (1 x p) row vector indicating which group of the G groups the p covariates in X belong to.}

\item{alpha_inits}{A (k x 1) column vector containing initial values for the regression coefficients corresponding to C.}

\item{beta_inits}{A (p x 1) column vector containing initial values for the regression coefficients corresponding to X.}

\item{lambda_sq_inits}{A (p x 1) column vector containing initial values for the local shrinkage parameters.}

\item{gamma_sq_inits}{A (G x 1) column vector containing initial values for the group shrinkage parameters.}

\item{a}{A (G x 1) column vector of shape parameters for the prior on the group shrinkage parameters.}

\item{b}{A (G x 1) column vector of shape parameters for the prior on the individual shrinkage parameters.}

\item{tau_sq_init}{Initial value for the global shrinkage parameter (double).}

\item{sigma_sq_init}{Initial value for the residual variance (double).}

\item{nu_init}{Initial value for the augmentation variable (double).}

\item{n_burn_in}{The number of burn-in samples (integer).}

\item{n_samples}{The number of posterior draws (integer).}

\item{n_thin}{The thinning interval (integer).}

\item{stable_const}{Parameter that controls numerical stability of the algorithm (double).}

\item{verbose}{Boolean value which indicates whether or not to print the progress of the Gibbs sampler.}

\item{btrick}{Boolean value which indicates whether or not to use the computational trick in Bhattacharya et al. (2016). Only recommended if number of covariates is much larger than the number of observations.}
}
\value{
A list containing the posterior draws of (1) the regression coefficients (alphas and betas) (2) the individual shrinkage parameters (lambda_sqs) (3) the group shrinkage parameters (gamma_sqs) (4) the global shrinkage parameter (tau_sqs) and (5) the residual error variance (sigma_sqs). The list also contains details regarding the dataset (X, C, Y, grp_idx) and Gibbs sampler details (n_burn_in, n_samples, and n_thin).
}
\description{
GIGG regression Gibbs sampler with fixed hyperparameters.
}
\examples{
X = concentrated$X
C = concentrated$C
Y = concentrated$Y
grp_idx = concentrated$grps
alpha_inits = concentrated$alpha
beta_inits = concentrated$beta

gf = gigg_fixed(X, C, Y, grp_idx, alpha_inits, beta_inits, lambda_sq_inits = rep(1, ncol(X)),
                gamma_sq_inits = rep(1, length(unique(grp_idx))),
                a = rep(0.5, length(unique(grp_idx))),
                b = rep(0.5, length(unique(grp_idx))),
                tau_sq_init = 1, sigma_sq_init = 1, nu_init = 1, n_burn_in = 500,
                n_samples = 1000, n_thin = 1, stable_const = 1e-07, verbose = TRUE,
                btrick = FALSE, stable_solve = FALSE)
}
