% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quick_solve}
\alias{quick_solve}
\title{Iterative one rank update for matrix inverse.}
\usage{
quick_solve(XtX_inv, D_pos, vec_draw)
}
\arguments{
\item{XtX_inv}{A precomputed (M x M) matrix inverse.}

\item{D_pos}{A (M x 1) vector of the square root of the diagonal entries in the D matrix.}

\item{vec_draw}{A (M x 1) vector drawn from a multivariate normal distribution.}
}
\value{
The solution to (XtX + D)*U = vec_draw.
}
\description{
An Rcpp function that computes the matrix inverse of XtX + D_pos.
}
